requirements.txt
Django>=4.0,<5.0
djangorestframework
djangorestframework-simplejwt
drf-yasg
pytest
pytest-django
psycopg2-binary
gunicorn
whitenoise


